Database Models Changed
The database model has been changed to allow the projects to be excluded from the final reports. There has been a skip value for projects which maybe un clear what to do with the data just yet.

Some work has been started to add a cli commands to the scripts.